+++ tests/test_genmf.py
@@ -41,14 +41,16 @@ To be more explicit, the power spectrum in all cases is produced with the follow
          'transfer__k_per_logint': 0,
          'transfer__kmax':100.0
 '''
+from itertools import product
+
 # ===============================================================================
 # Some Imports
 # ===============================================================================
 import numpy as np
-from hmf import MassFunction
-from astropy.cosmology import LambdaCDM
 import pytest
-from itertools import product
+from astropy.cosmology import LambdaCDM
+
+from hmf import MassFunction
 
 
 # =======================================================================
@@ -84,46 +86,51 @@ def max_diff(vec1, vec2, tol):
 # ===============================================================================
 # The Test Classes
 # ===============================================================================
-class TestGenMF(object):
-    def setup_method(self, test_method):
-        self.hmf = MassFunction(Mmin=7, Mmax=15.001, dlog10m=0.01,
-                                sigma_8=0.8, n=1,
-                                cosmo_model=LambdaCDM(Ob0=0.05, Om0=0.3, Ode0=0.7, H0=70.0, Tcmb0=0),
-                                lnk_min=-11, lnk_max=11, dlnk=0.01,
-                                transfer_params={"fname": "tests/data/transfer_for_hmf_tests.dat"},
-                                hmf_model='ST', z=0.0, transfer_model="FromFile", growth_model="GenMFGrowth")
+hmf = MassFunction(
+    Mmin=7, Mmax=15.001, dlog10m=0.01,
+    sigma_8=0.8, n=1,
+    cosmo_model=LambdaCDM(Ob0=0.05, Om0=0.3, Ode0=0.7, H0=70.0, Tcmb0=0),
+    lnk_min=-11, lnk_max=11, dlnk=0.01,
+    transfer_params={"fname": "tests/data/transfer_for_hmf_tests.dat"},
+    hmf_model='ST', z=0.0, transfer_model="FromFile", growth_model="GenMFGrowth",
+)
 
-    @staticmethod
-    def check_col(pert, fit, redshift, col):
-        """ Able to check all columns"""
-        data = np.genfromtxt("tests/data/" + fit + '_' + str(int(redshift)))[::-1][400:1201]
 
-        # We have to do funky stuff to the data if its been cut by genmf
-        if col is "sigma":
-            assert max_diff_rel(pert.sigma, data[:, 5], 0.004)
-        elif col is "lnsigma":  # We just do diff on this one because it passes through 0
-            assert max_diff(pert.lnsigma, data[:, 3], 0.001)
-        elif col is "n_eff":
-            assert max_diff_rel(pert.n_eff, data[:, 6], 0.001)
-        elif col is "dndlog10m":
-            assert rms_diff(pert.dndlog10m, 10 ** data[:, 1], 0.004)
-        elif col is "fsigma":
-            assert rms_diff(pert.fsigma, data[:, 4], 0.004)
-        elif col is "ngtm":
-            # The reason this is only good to 5% is GENMF's problem -- it uses
-            # poor integration.
-            assert rms_diff(pert.ngtm, 10 ** data[:, 2], 0.047)
+def check_col(pert, fit, redshift, col):
+    """ Able to check all columns"""
+    data = np.genfromtxt("tests/data/" + fit + '_' + str(int(redshift)))[::-1][400:1201]
 
-    @pytest.mark.parametrize(['redshift', 'col'],
-                             [(0.0, 'sigma'), (0.0, 'lnsigma'), (0.0, 'n_eff'),
-                              (2.0, 'sigma'), (2.0, 'lnsigma'), (2.0, 'n_eff')])
-    def test_sigmas(self, redshift, col):
-        # # Test z=0,2. Higher redshifts are poor in genmf.
-        self.hmf.update(z=redshift)
-        self.check_col(self.hmf, "ST", redshift, col)
+    # We have to do funky stuff to the data if its been cut by genmf
+    if col is "sigma":
+        assert max_diff_rel(pert.sigma, data[:, 5], 0.004)
+    elif col is "lnsigma":  # We just do diff on this one because it passes through 0
+        assert max_diff(pert.lnsigma, data[:, 3], 0.001)
+    elif col is "n_eff":
+        assert max_diff_rel(pert.n_eff, data[:, 6], 0.001)
+    elif col is "dndlog10m":
+        assert rms_diff(pert.dndlog10m, 10 ** data[:, 1], 0.004)
+    elif col is "fsigma":
+        assert rms_diff(pert.fsigma, data[:, 4], 0.004)
+    elif col is "ngtm":
+        # The reason this is only good to 5% is GENMF's problem -- it uses
+        # poor integration.
 
-    @pytest.mark.parametrize(['redshift', 'fit', 'col'],
-                             product([0.0,2.0], ["ST", "PS", "Reed03", "Warren", "Jenkins", "Reed07"], ['dndlog10m', 'ngtm', 'fsigma']))
-    def test_fits(self, redshift, fit, col):
-        self.hmf.update(z=redshift, hmf_model=fit)
-        self.check_col(self.hmf, fit, redshift, col)
+        assert rms_diff(pert.ngtm, 10 ** data[:, 2], 0.047)
+
+
+@pytest.mark.parametrize(['redshift', 'col'],
+                         [(0.0, 'sigma'), (0.0, 'lnsigma'), (0.0, 'n_eff'),
+                          (2.0, 'sigma'), (2.0, 'lnsigma'), (2.0, 'n_eff')])
+def test_sigmas(redshift, col):
+    # # Test z=0,2. Higher redshifts are poor in genmf.
+    hmf.update(z=redshift)
+    check_col(hmf, "ST", redshift, col)
+
+
+@pytest.mark.parametrize(
+    ['redshift', 'fit', 'col'],
+    product([0.0, 2.0], ["ST", "PS", "Reed03", "Warren", "Jenkins", "Reed07"], ['dndlog10m', 'ngtm', 'fsigma'])
+)
+def test_fits(redshift, fit, col):
+    hmf.update(z=redshift, hmf_model=fit)
+    check_col(hmf, fit, redshift, col)