+++ tests/test_integrate_hmf.py
@@ -19,30 +19,33 @@ def gammainc(z, x):
 
 
 class TestAnalyticIntegral(object):
-    def tggd(self, m, loghs, alpha, beta):
+    @staticmethod
+    def tggd(m, loghs, alpha, beta):
         return beta * (m / 10 ** loghs) ** alpha * np.exp(-(m / 10 ** loghs) ** beta)
 
-    def anl_int(self, m, loghs, alpha, beta):
+    @staticmethod
+    def anl_int(m, loghs, alpha, beta):
         return 10 ** loghs * gammainc((alpha + 1) / beta, (m / 10 ** loghs) ** beta)
 
-    def anl_m_int(self, m, loghs, alpha, beta):
-        return 10 ** (2 * loghs) * gammainc((alpha + 2) / beta, (m / 10 ** loghs) ** beta)
+    def test_basic(self):
+        m = np.logspace(10, 18, 500)
+        dndm = self.tggd(m, 14.0, -1.9, 0.8)
+        ngtm = self.anl_int(m, 14.0, -1.9, 0.8)
 
-    # def test_basic(self):
-    #     m = np.logspace(10,18,500)
-    #     dndm = self.tggd(m,14.0,-1.9,0.8)
-    #     ngtm = self.anl_int(m,14.0,-1.9,0.8)
-    #
-    #     print ngtm/hmf_integral_gtm(m,dndm)
-    #     assert np.allclose(ngtm,hmf_integral_gtm(m,dndm),rtol=0.03)
-    #
-    # def test_basic_mgtm(self):
-    #     m = np.logspace(10,18,500)
-    #     dndm = self.tggd(m,14.0,-1.9,0.8)
-    #     ngtm = self.anl_m_int(m,14.0,-1.9,0.8)
-    #
-    #     print ngtm/hmf_integral_gtm(m,dndm,True)
-    #     assert np.allclose(ngtm,hmf_integral_gtm(m,dndm,True),rtol=0.03)
+        full_numerical_ngtm = hmf_integral_gtm(m, dndm)
+
+        print(ngtm[m<1e15] / full_numerical_ngtm[m<1e15])
+        # Test integrating the whole integral at once, but only comparing
+        # smallish values
+        assert np.allclose(ngtm[m<1e15], full_numerical_ngtm[m<1e15], atol=0, rtol=0.03)
+
+        # more like what actually happens in hmf:
+        upper_integral = self.anl_int(m[m<1e15][-1], 14.0, -1.9, 0.8)
+
+        full_numerical_ngtm = hmf_integral_gtm(m[m<1e15], dndm[m<1e15]) + upper_integral
+
+        print(ngtm[m<1e15] / full_numerical_ngtm)
+        assert np.allclose(ngtm[m < 1e15], full_numerical_ngtm, atol=0, rtol=0.03)
 
     def test_high_z(self):
         m = np.logspace(10, 18, 500)
@@ -50,15 +53,17 @@ class TestAnalyticIntegral(object):
         ngtm = self.anl_int(m, 9.0, -1.93, 0.4)
 
         print(ngtm / hmf_integral_gtm(m, dndm))
-        assert np.allclose(ngtm, hmf_integral_gtm(m, dndm), rtol=0.03)
+        assert np.allclose(ngtm, hmf_integral_gtm(m, dndlnm=m * dndm), rtol=0.03)
 
-    # def test_low_mmax_z0(self):
-    #     m = np.logspace(10,15,500)
-    #     dndm = self.tggd(m,14.0,-1.9,0.8)
-    #     ngtm = self.anl_int(m,14.0,-1.9,0.8)
-    #
-    #     print ngtm/hmf_integral_gtm(m,dndm)
-    #     assert np.allclose(ngtm,hmf_integral_gtm(m,dndm),rtol=0.03)
+    def test_low_mmax_z0(self):
+        m = np.logspace(10,15,500)
+        dndm = self.tggd(m,14.0,-1.9,0.8)
+        ngtm = self.anl_int(m,14.0,-1.9,0.8)
+
+        upper_integral = self.anl_int(1e15, 14.0, -1.9, 0.8)
+        num_intg = hmf_integral_gtm(m,dndm, upper_integral=upper_integral)
+        print(ngtm/num_intg)
+        assert np.allclose(ngtm, num_intg, atol=0, rtol=0.03)
 
     def test_low_mmax_high_z(self):
         m = np.logspace(10, 15, 500)